+++ dagster_ray/kuberay/resources/__init__.py
@@ -1,0 +1,16 @@
+from dagster_ray.kuberay.resources.raycluster import (
+    KubeRayCluster,
+    RayClusterClient,
+    RayClusterClientResource,
+    RayClusterConfig,
+)
+from dagster_ray.kuberay.resources.rayjob import KubeRayJobClientResource, RayJobClient
+
+__all__ = [
+    "KubeRayCluster",
+    "RayClusterClientResource",
+    "RayClusterClient",
+    "RayClusterConfig",
+    "KubeRayJobClientResource",
+    "RayJobClient",
+]
