+++ pyproject.toml
@@ -70,6 +70,7 @@ dev = [
     "dunamai>=1.22.0",
     "pip>=24.2",
     "ipython>=8.12.3",
+    "syrupy>=4.9.1",
 ]
 
 [tool.hatch.version]
