+++ tests/kuberay/test_utils.py
@@ -1,8 +1,8 @@
 from dagster_ray.kuberay.utils import normalize_k8s_label_values
 
 
-def test_normalize_k8s_label_values():
-    assert normalize_k8s_label_values(
+def test_normalize_k8s_label_values(snapshot):
+    assert snapshot == normalize_k8s_label_values(
         {
             "foo": "bar",
             "my/label": "my/value",
@@ -13,13 +13,8 @@ def test_normalize_k8s_label_values():
             "badstart": "-foo",
             "badstart_after_initial_replace": "@foo",
         }
-    ) == {
-        "foo": "bar",
-        "my/label": "myvalue",
-        "user": "daniel-my-org",
-        "user-dirty": "daniel-my-org",
-        "alphanumeric": "abc123",
-        "long": 63 * "a",
-        "badstart": "foo",
-        "badstart_after_initial_replace": "foo",
-    }
+    )
+
+
+def test_normalize_k8s_label_values_important_labels():
+    assert normalize_k8s_label_values({"dagster/run-id": "12345"}) == {"dagster.io/run-id": "12345"}
